Web Dev/DSP/my-app/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
redis==5.3.1
yt-dlp==2025.3.31
ffmpeg-python==0.2.0
faster-whisper==1.1.1
argostranslate==1.9.6
webvtt-py==0.5.1
srt==3.5.3
TTS==0.22.0
